+++ angular_compiler/lib/angular_compiler.dart
@@ -2,7 +2,6 @@ export 'src/analyzer.dart'
     show
         getTypeName,
         linkToReference,
-        linkToReferenceDeprecated,
         linkTypeOf,
         typeArgumentOf,
         urlOf,