+++ angular/lib/src/compiler/view_compiler/compile_view.dart
@@ -4,7 +4,7 @@ import 'package:angular/src/core/change_detection/change_detection.dart'
 import 'package:angular/src/core/linker/view_type.dart' show ViewType;
 import "package:angular/src/core/metadata/view.dart" show ViewEncapsulation;
 import 'package:angular/src/facade/exceptions.dart' show BaseException;
-import 'package:angular/src/transform/common/names.dart'
+import 'package:angular/src/source_gen/common/names.dart'
     show toTemplateExtension;
 import 'package:angular_compiler/angular_compiler.dart';