+++ angular/lib/src/core/linker/element_ref.dart
@@ -2,15 +2,13 @@
 ///
 /// An `ElementRef` is backed by a render-specific element. In the browser, this
 /// is usually a DOM element.
-// Note: We don't expose things like `Injector`, `ViewContainer`, ... here, i.e.
-// users have to ask for what they need. With that, we can build better analysis
-// tools and could do better codegen in the future.
+
+/// **DEPRECATED**: A wrapper around a native DOM element inside of a View.
+///
+/// Inject `Element` or `HtmlElement` from `dart:html` instead; this will be
+/// removed in a future version of AngularDart, and has unnecessary overhead.
+@Deprecated('Inject or reference dart:html Element or HtmlElement instead')
 class ElementRef {
-  /// The underlying native element.
-  ///
-  /// Use with caution: Use this API as the last resort when direct access to
-  /// DOM is needed. Use templating and data-binding provided by Angular
-  /// instead.
   final dynamic nativeElement;
-  ElementRef(this.nativeElement);
+  const ElementRef(this.nativeElement);
 }